Multipart upload optimization.
In previous implementation, each parts is stored in temporary storage
then merged on multipart complete.
This creates double write (1. to temp storage. 2. when merging) which
really hits the performance.

In this commit, we only merged the metadata, not the data.
Which make it much faster and more reliable because we only need to do
single data write.

Also improve multipart test.